Core Dump on ec_throw(...)
As an alternative to a backtrace it may be more useful to generate a
core dump. The core dump has the advantage of containing a lot more
information and allowing inspection via a suitable debugger. The method
chosen here calls fork() and then immediately abort() on the child. This
permits the original process to continue on (possibly even recovering).
The disadvantage is that fork() is POSIX only.